requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-rest-knox==4.2.1
cryptography==42.0.8
python-decouple==3.8
